Make wrtc dependency optional (fixes #1)
wrtc brings along its own clang, and that version is outdated, and does not build on OSX and travis by default.

Unfortunately, the node.js wrtc module [does not seem to be maintained any more](https://github.com/js-platform/node-webrtc/commits/develop), and there is no alternative implementation for WebRTC.
In the long term, we can simply run these tests inside a webbrowser.
For now, make the dependency optional. If it fails building, we just continue.