order: change formula for closeness to league regulations
Formerly, we looked at the sum of the distance between ideal and actual position for each match.
However, it turns out this does not match human intuition that well. If a match from the end is moved forwards, then the rest should stay in order, but not at their former position.
Instead, we look at the number of pairwise incorrect orders.