courtspot/bupabfrage: Do not close connection
I misinterpreted the problem in Wesel. There are many connections in TIME_WAIT state (e.g. we and MySQL closed the connection, but the TCP 4-tuple is blocked ).

Closing connections is done automatically already, and does not help.

Instead, Windows CourtSpot should prefix the database host with `p:` to use persistent connections, like this:
```
/* Windows: */
	$DB_adress = 'p:127.0.0.1';
```