courtspot: Relay '' when exact player positioning is unknown
At the start of a match or game, we don't know the positions of the players. We used to leak implementation details, but fix this to properly relay an empty string, which gets turned into `null`.
This is visible to users; the displays won't show the second player serving anymore.